Add files via upload
Introducing an exciting project that combines the power of data scraping and music streaming to curate personalized playlists filled with nostalgia and joy!

This innovative tool begins by scraping the Billboard website, meticulously searching for the top 100 songs during the dates you input. Whether you're reminiscing about a milestone birthday, a memorable summer, or a significant event, our project ensures that the soundtrack of your past comes to life.

Once it compiles this treasure trove of musical memories, the magic continues with the integration of Spotipy, a powerful Spotify API. Seamlessly, it creates a custom playlist, tailored to your specific dates, featuring the very songs that once dominated the airwaves and your heart.

With just a few clicks, you're transported back in time, surrounded by the familiar melodies that accompanied your happiest moments. Whether you want to dance to the tunes of your youth or simply bask in the nostalgia of days gone by, our project is your ticket to a musical journey through time. Get ready to press play and let the memories flood back as you relive the magic of your past through the power of music.